tests/wpt/tests/css/css-grid/grid-lanes/tentative/parsing/item-tolerance-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masonry: item-tolerance parsing</title>
<link rel="help" href="https://drafts.csswg.org/css-grid-3">
<link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
<meta name="assert" content="item-tolerance supports the full grammar 'normal | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("item-tolerance", "normal");
test_valid_value("item-tolerance", "infinite");
test_valid_value("item-tolerance", "initial");

test_valid_value("item-tolerance", "0", "0px");
test_valid_value("item-tolerance", "1px");
test_valid_value("item-tolerance", "calc(2em + 3ex)");
test_valid_value("item-tolerance", "4%");
test_valid_value("item-tolerance", "5vmin");
</script>
</body>
</html>
